+++ CMakeLists.txt
@@ -6,4 +6,16 @@ if (NOT CMAKE_BUILD_TYPE)
     set(CMAKE_BUILD_TYPE Debug)
 endif()
 
-add_executable(main main.cpp)
+file(GLOB SRC
+    "*.cpp"
+)
+
+file(GLOB HEADER
+    "*.h"
+    "*.hpp"
+)
+
+add_executable(main ${SRC} ${HEADER})
+
+source_group("Source Files"            FILES ${SRC})
+source_group("Header Files"            FILES ${HEADER})
